requirements.txt
anyio==3.5.0
asgiref==3.5.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.3.2
fastapi==0.76.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pip==21.1.2
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.8
pytest-cov==3.0.0
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
setuptools==57.0.0
sniffio==1.2.0
starlette==0.18.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
wheel==0.36.2

pyserial~=3.5